+++ lib/dataset/test/data_array_test.cpp
@@ -66,6 +66,14 @@ TEST_F(DataArrayTest, name) {
   EXPECT_EQ(array.name(), "newname");
 }
 
+TEST_F(DataArrayTest, get_coord) {
+  DataArray a(data);
+  a.coords().set(Dim::X, coord);
+  EXPECT_EQ(a.coords().at(Dim::X), coord);
+  EXPECT_THROW_DISCARD(a.attrs().at(Dim::X), except::NotFoundError);
+  EXPECT_THROW_DISCARD(a.coords().at(Dim::Y), except::NotFoundError);
+}
+
 TEST_F(DataArrayTest, erase_coord) {
   DataArray a(data);
   a.coords().set(Dim::X, coord);